frontend: Charts: Refactor charts.tsx to Barrel Export Pattern.
