+++ frontend/src/components/cluster/Charts/index.ts
@@ -1,0 +1,4 @@
+import { CpuCircularChart, MemoryCircularChart } from './ResourceCharts';
+import { NodesStatusCircleChart, PodsStatusCircleChart } from './StatusCharts';
+
+export { MemoryCircularChart, CpuCircularChart, PodsStatusCircleChart, NodesStatusCircleChart };
